Fixed Choice Removes Quickly
I have gained feedback that when my app is used the choice that is picked shows up correctly in the modal. However when closing the modal the choice is the first thing to be reset just before the view is closed. I have added a timeout method to make sure the state isn't updated until after the modal screen has closed.